split ComputeCommitment into a getter and a computing func
This change is intended to prepare for the existence of two separate types of
Insert functions: those who will update on insertion and those who will
not. Calling Commit on the later is needed to compute the commitment at
the end of a batch insertion (typically, at the end of a block). Then, the
commitment can be read by calling Commitment() in both cases.