Use serialized commitments in HashedNode (#302)
* hashednode: update hashednode struct type

Computing the commitment is expensive and should only be done as needed.
This removes the hash *Fr field from the HashedNode struct type and updates
the commitment field from *Point to a []byte.

* hashednode: update deserialization to accomodate stateless tests

Node deserialization [parseNode()] was updated to return an
InternalNode. StatelessNode tests expect [parseNode()] to return
StatelessNode, so [parseStatelessNode()] function was added to
accomodate StatelessNode tests.

* stateless: remove unused toInternalNode function

* hashednode: simplify slice expressions to slice value itself

* hashednode: update commit method

Use SetBytesTrusted() for increased performance.

* stateless: remove unnecessary var declaration

* stateless: add stateless leaf deserialization

ParseStatelessNode() as implemented did not include functionality to
deserialize leaf nodes. Updated to fix.

Co-authored-by: Brandon Shiluk <b@shiluk.me>