GetProofItems now get the proof values from the tree
A remnant of the time that the prestate values were taken from
the witness, the data to be passed to the proof serialization code
was taken from it. But since the witness does not currently hold
the initial values, and since the data is taken from the tree, this
commit changes the interface for simplification.

This is enshrining a slower way to do things, but given the complexity
of the previous version, it's wiser to keep things simple at first,
and worry about optimization later.